Move and split test utils
