+++ saleor/graphql/product/tests/test_digital_content.py
@@ -2,8 +2,8 @@ import graphene
 
 from saleor.product.error_codes import ProductErrorCode
 from saleor.product.models import DigitalContent, ProductVariant
+from saleor.product.tests.utils import create_image
 from tests.api.utils import get_graphql_content, get_multipart_request_body
-from tests.utils import create_image
 
 
 def test_fetch_all_digital_contents(
